feat(admin): 2261 - Centres - Mise à jour selecteur de cohort (#3775)
* feat(admin): 2261 - Centres - Mise à jour selecteur de cohort (split)

* ajout du nouveau select de cohort

* fix design

* fix cohorlist info

* use SelectCohort
