+++ admin/package.json
@@ -28,6 +28,7 @@
     "@sentry/react": "^7.58.1",
     "@snu/ds": "*",
     "@swc/plugin-styled-components": "1.5.111",
+    "@tanstack/react-query": "^5.29.2",
     "bootstrap": "4.6.2",
     "browser-image-resizer": "^2.4.1",
     "chart.js": "^4.2.1",
