+++ admin/src/services/react-query.js
@@ -1,0 +1,6 @@
+import { capture } from "@/sentry";
+import { QueryCache, QueryClient } from "@tanstack/react-query";
+
+export const queryClient = new QueryClient({
+  queryCache: new QueryCache({ onError: (error) => capture(error) }),
+});
