+++ app/package.json
@@ -39,7 +39,7 @@
     "@sentry/react": "^7.58.1",
     "@sentry/vite-plugin": "^2.2.0",
     "@snu/ds": "*",
-    "@tanstack/react-query": "^5.29.0",
+    "@tanstack/react-query": "^5.29.2",
     "@uidotdev/usehooks": "^2.4.1",
     "autoprefixer": "^10.4.13",
     "bootstrap": "4.6.2",
